dashboard/requirements.txt
pandas>=2.0.0
pywebview>=4.4
